Merge upstream

+++ datadog/resource_datadog_integration_pagerduty.go
@@ -2,6 +2,7 @@ package datadog
 
 import (
 	"fmt"
+	"strings"
 	"sync"
 
 	"github.com/hashicorp/terraform/helper/schema"
@@ -16,6 +17,7 @@ func resourceDatadogIntegrationPagerduty() *schema.Resource {
 	return &schema.Resource{
 		Create: resourceDatadogIntegrationPagerdutyCreate,
 		Read:   resourceDatadogIntegrationPagerdutyRead,
+		Exists: resourceDatadogIntegrationPagerdutyExists,
 		Update: resourceDatadogIntegrationPagerdutyUpdate,
 		Delete: resourceDatadogIntegrationPagerdutyDelete,
 		Importer: &schema.ResourceImporter{
@@ -150,6 +152,20 @@ func resourceDatadogIntegrationPagerdutyRead(d *schema.ResourceData, meta interf
 	return nil
 }
 
+func resourceDatadogIntegrationPagerdutyExists(d *schema.ResourceData, meta interface{}) (b bool, e error) {
+	client := meta.(*datadog.Client)
+
+	_, err := client.GetIntegrationPD()
+	if err != nil {
+		if strings.Contains(err.Error(), "404 Not Found") {
+			return false, nil
+		}
+		return false, err
+	}
+
+	return true, nil
+}
+
 func resourceDatadogIntegrationPagerdutyUpdate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*datadog.Client)
 	integrationPdMutex.Lock()
@@ -167,15 +183,17 @@ func resourceDatadogIntegrationPagerdutyUpdate(d *schema.ResourceData, meta inte
 	// if there are none currently configured services, we actually
 	// have to remove them explicitly, otherwise the underlying API client
 	// would not send the "services" key at all and they wouldn't get deleted
-	currentServices := d.Get("services").([]interface{})
-	if len(currentServices) == 0 {
-		pd, err := client.GetIntegrationPD()
-		if err != nil {
-			return fmt.Errorf("Error while deleting Pagerduty integration service object: %v", err)
-		}
-		for _, service := range pd.Services {
-			if err := client.DeleteIntegrationPDService(*service.ServiceName); err != nil {
+	if value, ok := d.GetOk("individual_services"); !ok || !value.(bool) {
+		currentServices := d.Get("services").([]interface{})
+		if len(currentServices) == 0 {
+			pd, err := client.GetIntegrationPD()
+			if err != nil {
 				return fmt.Errorf("Error while deleting Pagerduty integration service object: %v", err)
+			}
+			for _, service := range pd.Services {
+				if err := client.DeleteIntegrationPDService(*service.ServiceName); err != nil {
+					return fmt.Errorf("Error while deleting Pagerduty integration service object: %v", err)
+				}
 			}
 		}
 	}
